Fixed #161, DHDs now work near the world border

+++ src/main/java/mrjake/aunis/raycaster/util/DHDVertex.java
@@ -12,21 +12,21 @@ public class DHDVertex {
 	public final float y;
 	public final float z;
 	
-	private float xRotated;
-	private float yRotated;
-	private float zRotated;
+	private double xRotated;
+	private double yRotated;
+	private double zRotated;
 	
-	private float xGlobal;
-	private float yGlobal;
-	private float zGlobal;
+	private double xGlobal;
+	private double yGlobal;
+	private double zGlobal;
 	
-	public float getGlobalX() { return xGlobal; }
-	public float getGlobalY() { return yGlobal; }
-	public float getGlobalZ() { return zGlobal; }
+	public double getGlobalX() { return xGlobal; }
+	public double getGlobalY() { return yGlobal; }
+	public double getGlobalZ() { return zGlobal; }
 	
-	public float xDiffrence;
-	public float yDiffrence;
-	public float zDiffrence;
+	public double xDiffrence;
+	public double yDiffrence;
+	public double zDiffrence;
 	
 	BlockPos oldGlobal = null;
 	private Vec3d lastPlayerPos = new Vec3d(0,0,0);
